[Move] optimizeImage 파일 이동 및 파일명 수정
src/constants/OptimizeImage.ts ->  src/utils/optimizeImage.ts

1. 상수가 아니라 함수이므로 constants 에서 utils 로 이동
2. 컴포넌트가 아니므로 파일명 첫 글자를 소문자로 작성
