+++ src/hooks/queries/feed.query.ts
@@ -20,7 +20,7 @@ import {
   Comment,
   uploadImage,
 } from '@/apis/feed.api';
-import { optimizeImage } from '@/constants/OptimizeImage';
+import { optimizeImage } from '@/utils/optimizeImage ';
 import { querySuccessHandler } from '@/utils/querySuccessHandler';
 import {
   InfiniteData,
